backend/pyproject.toml
[project]
name = "hera-agent"
version = "0.0.1"
description = "Agentic motion-graphics generator on top of Hera, built with LangGraph."
requires-python = ">=3.12"
dependencies = [
    # Web / API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.1",
    "httpx>=0.27",
    "python-multipart>=0.0.20",
    # Data / config
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    # Database (Neon Postgres + pgvector)
    "psycopg[binary,pool]>=3.2",
    "pgvector>=0.3.6",
    # LangGraph
    "langgraph>=0.2.50",
    "langgraph-checkpoint-postgres>=2.0",
    # Gemini
    "google-genai>=0.3",
    # Article extraction fallback
    "readability-lxml>=0.8.1",
    "lxml>=5.3",
    "beautifulsoup4>=4.12",
    # Misc
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "asgi-lifespan>=2.1",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]
extend-safe-fixes = ["UP"]

[tool.ruff.lint.flake8-bugbear]
# FastAPI's Query/Path/Body/Depends/Header/Form are immutable sentinels — common in arg defaults.
extend-immutable-calls = [
    "fastapi.Query",
    "fastapi.Path",
    "fastapi.Body",
    "fastapi.Depends",
    "fastapi.Header",
    "fastapi.Form",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
